[#1216] neofs-cli: Reuse key retrieving code between modules
